routes: migrate models domain (15 routes)
Move to src/server/routes/models.ts:
- /api/models, /api/image-models
- /api/custom-providers (GET, POST, DELETE :id, /test)
- /api/provider-keys (GET, POST :provider, DELETE :provider)
- /api/aigw (status, configure POST/DELETE, test, refresh)
- /api/models/test
- /api/aigw/v1/* proxy

Also adds broadcastPreferencesChanged + getSafePreferences to cross-project.ts.

Net: server.ts -325 lines (7561 → 7236). No behaviour change.
