+++ src/server/server.ts
@@ -72,12 +72,10 @@ import { progressBus as searchProgressBus } from "./search/progress-bus.js";
 import { isSandboxAllowed } from "./auth/sandbox-guard.js";
 import * as previewMount from "./preview/mount.js";
 import { broadcastPreviewChanged, subscribePreviewChanged } from "./preview/events.js";
-import { configureAigw, removeAigw, getAigwUrl, discoverAigwModels, proxyRequest, startupAigwCheck, writeContextWindowOverrides } from "./agent/aigw-manager.js";
+import { startupAigwCheck, writeContextWindowOverrides } from "./agent/aigw-manager.js";
 import { writeOpenAIModelAdditions } from "./agent/openai-model-additions.js";
 import { ReviewAnnotationStore, type ReviewAnnotation } from "./review-annotation-store.js";
-import { getAvailableModels, discoverModelsForConfig, invalidateModelCache } from "./agent/model-registry.js";
-import type { CustomProviderConfig } from "./agent/model-registry.js";
-import { getAvailableImageModels } from "./agent/image-generation.js";
+
 import { ProjectRegistry, SymlinkProjectRootError } from "./agent/project-registry.js";
 import { ProjectContextManager } from "./agent/project-context-manager.js";
 import { resolveProjectForRequest } from "./agent/resolve-project.js";
@@ -3134,329 +3132,6 @@ async function handleApiRoute(
 	}
 
 	// ── Unified Model Registry ──
-
-	// GET /api/models — unified model list from all sources
-	if (url.pathname === "/api/models" && req.method === "GET") {
-		try {
-			const models = await getAvailableModels(preferencesStore);
-			json(models);
-		} catch (err: any) {
-			jsonError(500, err, { error: `Failed to load models: ${err.message}` });
-		}
-		return;
-	}
-
-	// GET /api/image-models — image generation model list
-	if (url.pathname === "/api/image-models" && req.method === "GET") {
-		try {
-			json(getAvailableImageModels(preferencesStore));
-		} catch (err: any) {
-			jsonError(500, err, { error: `Failed to load image models: ${err.message}` });
-		}
-		return;
-	}
-
-	// ── Custom Providers ──
-
-	// GET /api/custom-providers — list all custom provider configs
-	if (url.pathname === "/api/custom-providers" && req.method === "GET") {
-		const configs = (preferencesStore.get("customProviders") as CustomProviderConfig[] | undefined) || [];
-		json(configs);
-		return;
-	}
-
-	// POST /api/custom-providers/test — discover models without persisting
-	if (url.pathname === "/api/custom-providers/test" && req.method === "POST") {
-		const body = await readBody(req);
-		if (!body || !body.type || !body.baseUrl) {
-			json({ error: "Missing required fields: type, baseUrl" }, 400);
-			return;
-		}
-		const config: CustomProviderConfig = {
-			id: body.id || "test-" + Date.now(),
-			name: body.name || body.type,
-			type: body.type,
-			baseUrl: body.baseUrl,
-			...(body.apiKey ? { apiKey: body.apiKey } : {}),
-		};
-		try {
-			const models = await discoverModelsForConfig(config);
-			json({ models });
-		} catch (err: any) {
-			jsonError(500, err);
-		}
-		return;
-	}
-
-	// POST /api/custom-providers — add or update a custom provider config
-	if (url.pathname === "/api/custom-providers" && req.method === "POST") {
-		const body = await readBody(req);
-		if (!body || !body.id || !body.type || !body.baseUrl) {
-			json({ error: "Missing required fields: id, type, baseUrl" }, 400);
-			return;
-		}
-		const configs = (preferencesStore.get("customProviders") as CustomProviderConfig[] | undefined) || [];
-		const existing = configs.findIndex((c: CustomProviderConfig) => c.id === body.id);
-		const config: CustomProviderConfig = {
-			id: body.id,
-			name: body.name || body.id,
-			type: body.type,
-			baseUrl: body.baseUrl,
-			...(body.apiKey ? { apiKey: body.apiKey } : {}),
-			...(body.models ? { models: body.models } : {}),
-		};
-		if (existing >= 0) {
-			configs[existing] = config;
-		} else {
-			configs.push(config);
-		}
-		preferencesStore.set("customProviders", configs);
-		json({ ok: true, config });
-		return;
-	}
-
-	// DELETE /api/custom-providers/:id — remove a custom provider config
-	if (url.pathname.startsWith("/api/custom-providers/") && req.method === "DELETE") {
-		const providerId = decodeURIComponent(url.pathname.slice("/api/custom-providers/".length));
-		if (!providerId) {
-			json({ error: "Missing provider id" }, 400);
-			return;
-		}
-		const configs = (preferencesStore.get("customProviders") as CustomProviderConfig[] | undefined) || [];
-		const filtered = configs.filter((c: CustomProviderConfig) => c.id !== providerId);
-		preferencesStore.set("customProviders", filtered);
-		json({ ok: true });
-		return;
-	}
-
-	// ── Provider Keys ──
-
-	// GET /api/provider-keys — list providers that have keys set (no key values)
-	if (url.pathname === "/api/provider-keys" && req.method === "GET") {
-		const all = preferencesStore.getAll();
-		const providers = Object.keys(all)
-			.filter(k => k.startsWith("providerKey.") && all[k])
-			.map(k => k.slice("providerKey.".length));
-		json({ providers });
-		return;
-	}
-
-	// POST /api/provider-keys/:provider — store a provider API key
-	if (url.pathname.startsWith("/api/provider-keys/") && req.method === "POST") {
-		const provider = decodeURIComponent(url.pathname.slice("/api/provider-keys/".length));
-		if (!provider) {
-			json({ error: "Missing provider name" }, 400);
-			return;
-		}
-		const body = await readBody(req);
-		if (!body?.key || typeof body.key !== "string") {
-			json({ error: "Missing 'key' field" }, 400);
-			return;
-		}
-		preferencesStore.set(`providerKey.${provider}`, body.key);
-		json({ ok: true });
-		return;
-	}
-
-	// DELETE /api/provider-keys/:provider — remove a provider API key
-	if (url.pathname.startsWith("/api/provider-keys/") && req.method === "DELETE") {
-		const provider = decodeURIComponent(url.pathname.slice("/api/provider-keys/".length));
-		if (!provider) {
-			json({ error: "Missing provider name" }, 400);
-			return;
-		}
-		preferencesStore.remove(`providerKey.${provider}`);
-		json({ ok: true });
-		return;
-	}
-
-	// ── AI Gateway ──
-
-	// GET /api/aigw/status — check if aigw is configured
-	if (url.pathname === "/api/aigw/status" && req.method === "GET") {
-		const aigwUrl = getAigwUrl(preferencesStore);
-		if (!aigwUrl) {
-			json({ configured: false });
-		} else {
-			// Discover fresh models instead of reading from preferences cache
-			try {
-				const models = await discoverAigwModels(aigwUrl);
-				json({ configured: true, url: aigwUrl, models });
-			} catch {
-				json({ configured: true, url: aigwUrl, models: [] });
-			}
-		}
-		return;
-	}
-
-	// POST /api/aigw/configure — set aigw URL, discover models, write models.json
-	if (url.pathname === "/api/aigw/configure" && req.method === "POST") {
-		const body = await readBody(req);
-		if (!body?.url || typeof body.url !== "string") {
-			json({ error: "Missing 'url' field" }, 400);
-			return;
-		}
-		try {
-			const models = await configureAigw(body.url, preferencesStore);
-			invalidateModelCache();
-			broadcastPreferencesChanged();
-			json({ ok: true, models });
-		} catch (err: any) {
-			jsonError(502, err, { error: `Failed to configure AI Gateway: ${err.message}` });
-		}
-		return;
-	}
-
-	// DELETE /api/aigw/configure — remove aigw config
-	if (url.pathname === "/api/aigw/configure" && req.method === "DELETE") {
-		removeAigw(preferencesStore);
-		invalidateModelCache();
-		broadcastPreferencesChanged();
-		json({ ok: true });
-		return;
-	}
-
-	// POST /api/aigw/test — test connection to a URL without saving
-	if (url.pathname === "/api/aigw/test" && req.method === "POST") {
-		const body = await readBody(req);
-		if (!body?.url || typeof body.url !== "string") {
-			json({ error: "Missing 'url' field" }, 400);
-			return;
-		}
-		try {
-			const models = await discoverAigwModels(body.url);
-			json({ ok: true, models });
-		} catch (err: any) {
-			jsonError(502, err);
-		}
-		return;
-	}
-
-	// POST /api/aigw/refresh — re-discover models from the configured gateway
-	if (url.pathname === "/api/aigw/refresh" && req.method === "POST") {
-		const aigwUrl = getAigwUrl(preferencesStore);
-		if (!aigwUrl) {
-			json({ error: "No AI Gateway configured" }, 400);
-			return;
-		}
-		try {
-			const models = await configureAigw(aigwUrl, preferencesStore);
-			invalidateModelCache();
-			broadcastPreferencesChanged();
-			json({ models });
-		} catch (err: any) {
-			jsonError(502, err);
-		}
-		return;
-	}
-
-	// POST /api/models/test — send a trivial "Reply with OK" completion to verify
-	// that a default-model preference actually resolves and responds. Used by
-	// the Settings > Models tab per-row Test button.
-	if (url.pathname === "/api/models/test" && req.method === "POST") {
-		const body = await readBody(req);
-		const pref = typeof body?.pref === "string" ? body.pref.trim() : "";
-		if (!pref) {
-			json({ ok: false, error: "Missing 'pref' field" }, 400);
-			return;
-		}
-		const slash = pref.indexOf("/");
-		if (slash <= 0) {
-			json({ ok: false, error: "Malformed pref — expected 'provider/modelId'" }, 400);
-			return;
-		}
-		const provider = pref.slice(0, slash);
-		const modelId = pref.slice(slash + 1);
-		try {
-			const models = await getAvailableModels(preferencesStore);
-			const resolved = models.find((m) => m.provider === provider && m.id === modelId);
-			if (!resolved) {
-				json({
-					ok: false,
-					error: `Model "${pref}" is not in the current available-models list. It may be a stale preference.`,
-				}, 404);
-				return;
-			}
-			if (provider !== "aigw") {
-				json({
-					ok: false,
-					modelResolved: resolved.id,
-					error: "Test not supported for this provider yet — only AI Gateway models can be tested from here.",
-				}, 422);
-				return;
-			}
-			const aigwUrl = getAigwUrl(preferencesStore);
-			if (!aigwUrl) {
-				json({ ok: false, error: "No AI Gateway configured." });
-				return;
-			}
-			const baseUrl = aigwUrl.replace(/\/+$/, "");
-			const chatUrl = baseUrl.endsWith("/v1") ? `${baseUrl}/chat/completions` : `${baseUrl}/v1/chat/completions`;
-
-			// The aigw registry strips the provider prefix (e.g. "aws/") from Claude
-			// model IDs; reconstruct the full ID by querying the gateway's /v1/models.
-			let sendId = modelId;
-			try {
-				const modelsUrl = baseUrl.endsWith("/v1") ? `${baseUrl}/models` : `${baseUrl}/v1/models`;
-				const r = await fetch(modelsUrl, { signal: AbortSignal.timeout(5000) });
-				if (r.ok) {
-					const data = await r.json() as { data?: Array<{ id: string }> };
-					if (Array.isArray(data.data)) {
-						const exact = data.data.find((m) => m.id === modelId);
-						if (exact) sendId = exact.id;
-						else {
-							const match = data.data.find((m) => {
-								const idx = m.id.indexOf("/");
-								return idx >= 0 && m.id.slice(idx + 1) === modelId;
-							});
-							if (match) sendId = match.id;
-						}
-					}
-				}
-			} catch {
-				/* keep sendId = modelId — gateway will reject if wrong */
-			}
-			const started = Date.now();
-			try {
-				const resp = await fetch(chatUrl, {
-					method: "POST",
-					headers: { "Content-Type": "application/json" },
-					body: JSON.stringify({
-						model: sendId,
-						max_tokens: 5,
-						messages: [
-							{ role: "user", content: "Reply with OK" },
-						],
-					}),
-					signal: AbortSignal.timeout(15000),
-				});
-				const latencyMs = Date.now() - started;
-				if (!resp.ok) {
-					const errText = (await resp.text().catch(() => "")).slice(0, 300);
-					json({ ok: false, modelResolved: sendId, latencyMs, error: `Gateway ${resp.status}: ${errText || resp.statusText}` });
-					return;
-				}
-				// Best-effort parse; we don't require specific text content—just a successful round-trip.
-				await resp.json().catch(() => ({}));
-				json({ ok: true, modelResolved: sendId, latencyMs });
-			} catch (err: any) {
-				const latencyMs = Date.now() - started;
-				json({ ok: false, modelResolved: sendId, latencyMs, error: err?.message || "Request failed" });
-			}
-		} catch (err: any) {
-			jsonError(500, err, { ok: false, error: err?.message || "Test failed" });
-		}
-		return;
-	}
-
-	// Proxy: /api/aigw/v1/* → forward to configured aigw URL
-	if (url.pathname.startsWith("/api/aigw/v1/") && getAigwUrl(preferencesStore)) {
-		const aigwUrl = getAigwUrl(preferencesStore)!;
-		const subPath = url.pathname.replace("/api/aigw/v1/", "/v1/");
-		const targetUrl = `${aigwUrl}${subPath}${url.search}`;
-		proxyRequest(targetUrl, req, res);
-		return;
-	}
 
 	// GET /api/roles/assistant/prompts — must come before :name route
 	if (url.pathname === "/api/roles/assistant/prompts" && req.method === "GET") {
